Move TextCopyBox into shared components
