+++ Source/SelfService/Web/integrations/connection/configuration/IonServiceAccountCredentials.tsx
@@ -9,7 +9,7 @@ import { enqueueSnackbar } from 'notistack';
 import { FileUploadForm, FileUploadFormRef, MaxWidthTextBlock, MaxWidthBlock } from '@dolittle/design-system';
 import { IonConfigRequest } from '../../../apis/integrations/generated';
 
-import { TextCopyBox } from './TextCopyBox';
+import { TextCopyBox } from '../../../components/TextCopyBox';
 
 /**
  * Simple type guard to check if the object is of type IonConfigRequest.
